% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkg.R
\name{gkg}
\alias{gkg}
\title{Function to retrive Data from Google Knowledge Graph API as a Data Frame output}
\usage{
gkg(query, token = "AIzaSyDsxs2vKm33doRABnW4JoieK8RrDQvJwds",
  language = "en", limit = 1, types = "", prefix = FALSE,
  type_output = TRUE)
}
\description{
This function allows you to get the Information for a specific Keyword you enter. The output is a Dataframe with the Data retrived by the Google API.
}
