% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkg_raw.R
\name{gkg_raw}
\alias{gkg_raw}
\title{Function to retrive Data from Google Knowledge Graph API}
\usage{
gkg_raw(query, token = "AIzaSyDsxs2vKm33doRABnW4JoieK8RrDQvJwds",
  language = "en", limit = 1, types = "", prefix = FALSE)
}
\description{
This function allows you to get the Information for a specific Keyword you enter.
}
